debug

+++ src/zeroband/models.py
@@ -7,6 +7,8 @@ from transformers import (
     Qwen2ForCausalLM,
 )
 
+import torch
+
 ModelName: TypeAlias = Literal["debugmodel", "150M", "1B", "Qwen32B", "Qwen1.5B", "Qwen7B", "Llama8B", "QwQ32B"]
 ModelType: TypeAlias = LlamaForCausalLM | Qwen2ForCausalLM
 AttnImpl: TypeAlias = Literal["flex_attention", "sdpa", "flash_attention_2"]
@@ -50,6 +52,6 @@ def get_model_and_tokenizer(model_name: ModelName, attn_impl: AttnImpl) -> tuple
     tokenizer = AutoTokenizer.from_pretrained(name_to_hf_tokenizer[model_name])
     config_model = config_class.from_pretrained(name_to_hf_model[model_name], attn_implementation=attn_impl)
     config_model.use_cache = False
-    model = model_class.from_pretrained(pretrained_model_name_or_path=name_to_hf_model[model_name], config=config_model)
+    model = model_class.from_pretrained(pretrained_model_name_or_path=name_to_hf_model[model_name], config=config_model, torch_dtype=torch.bfloat16)
     tokenizer.pad_token_id = tokenizer.eos_token_id
     return model, tokenizer  # type: ignore
